backend2/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
imagekitio==2.2.5
psycopg2-binary==2.8.6
python-dotenv==0.15.0
django-mailgun==0.9.1
djangorestframework-camel-case==1.2.0
django-update-from-dict
